# SC: Add checking the already handled events for Value Transfer Recovery